// apps/provider/src/app/shared/utils/units/conversion.types.ts
export type Metric =
  | 'composition'
  | 'circumference'
  | 'distance'
  | 'duration'
  | 'height'
  | 'speed'
  | 'step'
  | 'cal'
  | 'volume'
  | 'temperature-push'
  | 'temperature-fetch'

export const POUND = 453.592 // gr
export const POUNDS = 0.00220462 // 1 gr

export const OUNCES = 0.033814 // 1 ml

export const INCH = 25.4 // mm
export const INCHES = 0.0393701 // 1 mm

export const FEETS = 0.00328084 // 1 mm

export const MILES = 0.000621371 // 1 m

export const KILOMETER = 1320 // steps
// 2000 steps ~ 1 mile